pyproject.toml
[project]
name = "pydanticai2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai[examples,logfire]>=0.0.15",
    "python-dotenv>=1.0.1",
]
